refactor: create shared SplitViewGrid for Benefits
