--- src/components/pages/ai/benefits/index.js
@@ -1,3 +1,0 @@
-import Benefits from './benefits';
-
-export default Benefits;
