+++ src/components/shared/split-view-grid/index.js
@@ -1,0 +1,3 @@
+import SplitViewGrid from './split-view-grid';
+
+export default SplitViewGrid;
